+++ core/kits/assets/css/panel.scss
@@ -64,13 +64,6 @@
 	}
 }
 
-#elementor-kit-panel-menu {
-
-	.elementor-panel-menu-items {
-		margin: 15px 0 0;
-	}
-}
-
 .elementor-control {
 
 	&-type-global-style-repeater {
