+++ core/kits/documents/kit.php
@@ -23,12 +23,17 @@ class Kit extends PageBase {
 		parent::__construct( $data );
 
 		$this->tabs = [
-			'site_identity' => new Tabs\Site_Identity( $this ),
-			'lightbox' => new Tabs\Lightbox( $this ),
 			'global-colors' => new Tabs\Global_Colors( $this ),
 			'global-typography' => new Tabs\Global_Typography( $this ),
-			'layout_settings' => new Tabs\Layout_Settings( $this ),
-			'theme_style' => new Tabs\Theme_Style( $this ),
+			'theme-style-typography' => new Tabs\Theme_Style_Typography( $this ),
+			'theme-style-buttons' => new Tabs\Theme_Style_Buttons( $this ),
+			'theme-style-images' => new Tabs\Theme_Style_Images( $this ),
+			'theme-style-form-fields' => new Tabs\Theme_Style_Form_Fields( $this ),
+			'settings-site-identity' => new Tabs\Settings_Site_Identity( $this ),
+			'settings-background' => new Tabs\Settings_Background( $this ),
+			'settings-layout' => new Tabs\Settings_Layout( $this ),
+			'settings-lightbox' => new Tabs\Settings_Lightbox( $this ),
+			'settings-custom-css' => new Tabs\Settings_Custom_CSS( $this ),
 		];
 	}
